+++ simulator/engine/index.js
@@ -1,5 +1,6 @@
 export { default as IdolConfig } from "./IdolConfig";
 export { default as StageConfig } from "./StageConfig";
+export { default as IdolStageConfig } from "./IdolStageConfig";
 export { default as StageEngine } from "./StageEngine";
 export { default as StageLogger } from "./StageLogger";
 export { default as StagePlayer } from "./StagePlayer";